moving some code around
